refactor(mcp): extract status tool module
