+++ openspec/changes/refactor-mcp-status-tool-module/README.md
@@ -1,0 +1,3 @@
+# refactor-mcp-status-tool-module
+
+Refactor: extract MCP status tool implementation into a dedicated module
